# swagger.yaml
swagger: "2.0"
info:
  description: GovReady provides a dashboard and tools to enhance security for government websites and achieve FISMA compliance.
  version: "1.0.0"
  title: GovReady API
  termsOfService: http://govready.com/tos
  contact:
    name: ops@govready.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: plugin.govready.com
basePath: /v1.0
schemes:
  - http
paths:
  /initialize:
    post:
      tags:
        - site
      summary: Initialize a new site
      description: ""
      operationId: initializeSite
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - in: query
          name: url
          description: The url (method and domain) of the site that is getting initialized
          required: true
          type: string
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: "#/definitions/Site"
        "405":
          description: Invalid input
  /token:
    post:
      tags:
        - site
      summary: Initialize a new site
      description: ""
      operationId: token
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: "#/definitions/Key"
      security:
        - govready_auth:
          - write_sites
          - read_sites
...
  /sites/{siteId}:
    get:
      tags:
        - site
      summary: Find Site by ID
      description: Returns a site when ID < 10.  ID > 10 or nonintegers will simulate API error conditions
      operationId: getsiteById
      produces:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: siteId
          description: ID of site that needs to be fetched
          required: true
          type: string
      responses:
        "404":
          description: site not found
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/Site"
        "400":
          description: Invalid ID supplied
      security:
        - govready_auth:
          - write_sites
          - read_sites
    post:
      tags:
        - site
      summary: Updates a site with form data
      description: ""
      operationId: updatesiteWithForm
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: siteId
          description: ID of site that needs to be updated
          required: true
          type: string
        - in: formData
          name: name
          description: Updated name of the site
          required: true
          type: string
        - in: formData
          name: status
          description: Updated status of the site
          required: true
          type: string
      responses:
        "405":
          description: Invalid input
      security:
        - govready_auth:
          - write_sites
          - read_sites
    delete:
      tags:
        - site
      summary: Deletes a site
      description: ""
      operationId: deletesite
      produces:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: siteId
          description: site id to delete
          required: true
          type: string
      responses:
        "400":
          description: Invalid site value
      security:
        - govready_auth:
          - write_sites
          - read_sites
...
  /users:
    post:
      tags:
        - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      produces:
        - application/json
        - application/xml
      parameters:
        - in: body
          name: body
          description: Created user object
          required: false
          schema:
            $ref: "#/definitions/User"
      responses:
        default:
          description: successful operation
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
  govready_auth:
    type: oauth2
    authorizationUrl: https://govready.auth0.com/authorize
    flow: implicit
    scopes:
      write_sites: modify sites in your account
      read_sites: read your sites
definitions:
  Key:
    type: object
    properties:
      key:
        type: string
        format: int64
        description: The API public key
      secret:
        type: string
        format: int64
        description: The api secret key
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      userStatus:
        type: integer
        format: int32
        description: User Status
  Site:
    type: object
    required:
      - name
      - url
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        example: GovReady
      url:
        type: string
        example: http://govready.com
        description: Full URL to the site homepage
      status:
        type: string
        example: active
        description: Site status
      application:
        $ref: "#/definitions/Application"
        description: The Application object this site is built with
...
  Application:
    type: object
    properties:
      namespace:
        type: string
        example: wordpress
      label:
        type: string
        example: WordPress
      version:
        description: string
        example: 4.4.2